pyproject.toml
[project]
name = "google-workspace-mcp-server"
version = "0.1.0"
description = "MCP server for Google Workspace API (Gmail, Sheets, Docs, Drive, Slides) with OAuth 2.0"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "mcp>=1.7.0",
    "google-api-python-client>=2.100.0",
    "google-auth-oauthlib>=1.0.0",
    "google-auth>=2.20.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.scripts]
google-workspace-mcp = "google_workspace_mcp.server:main"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]
